% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirRoute-package.R
\docType{package}
\name{AirRoute-package}
\alias{AirRoute}
\alias{AirRoute-package}
\title{AirRoute: k-d Tree Nearest Neighbours}
\description{
Fast k-nearest neighbour search with a pure C k-d tree core and R bindings. Training data is referenced without copying from R matrices. Part of the broader AirRoute project (future: Python bindings and geodesic route metrics).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BartekDomanowski/AirRoute}
  \item Report bugs at \url{https://github.com/BartekDomanowski/AirRoute/issues}
}

}
\author{
\strong{Maintainer}: Bartłomiej Domanowski \email{01187317@pw.edu.pl} [copyright holder]

}
\keyword{internal}
