% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{cross_track_route_km}
\alias{cross_track_route_km}
\title{cross-track distance to a polyline route}
\usage{
cross_track_route_km(lat_p, lon_p, plan_lats, plan_lons)
}
\arguments{
\item{lat_p, lon_p}{point P (degs).}

\item{plan_lats, plan_lons}{planned route waypoints (degs).}
}
\value{
minimum cross-track distance in km.
}
\description{
minimum cross-track distance from P to any segment of the planned route.
}
