% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{cross_track_segment_km}
\alias{cross_track_segment_km}
\title{cross-track distance to one route segment}
\usage{
cross_track_segment_km(lat_p, lon_p, lat_a, lon_a, lat_b, lon_b)
}
\arguments{
\item{lat_p, lon_p}{point P (degs).}

\item{lat_a, lon_a, lat_b, lon_b}{segment endpoints A and B (degs).}
}
\value{
cross-track distance in km.
}
\description{
perpendicular distance from point P to great-circle segment A--B (km).
}
