% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{dist_haversine_km}
\alias{dist_haversine_km}
\title{distance (Haversine)}
\usage{
dist_haversine_km(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude of point 1 (degs).}

\item{lon1}{longitude of point 1 (degs).}

\item{lat2}{latitude of point 2 (degs).}

\item{lon2}{longitude of point 2 (degs).}
}
\value{
distance in km.
}
\description{
computes the distance between two points on the WGS84 sphere
using the haversine formula (with Earth radius 6371.0088 km).
}
\examples{
dist_haversine_km(52.1657, 20.9671, 51.47, -0.4543)  # waw -> lhr ~ 1460-70km

}
