% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{initial_bearing_deg}
\alias{initial_bearing_deg}
\title{initial bearing (deg)}
\usage{
initial_bearing_deg(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude of point 1 (degs).}

\item{lon1}{longitude of point 1 (degs).}

\item{lat2}{latitude of point 2 (degs).}

\item{lon2}{longitude of point 2 (degs).}
}
\value{
initial bearing in degrees, clockwise from north (0--360).
}
\description{
computes the initial bearing (deg) between two points on the WGS84 sphere.
}
\examples{
initial_bearing_deg(52.1657, 20.9671, 50.0379, 8.5622)  # WAW -> FRA ~ 260 deg

}
