% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myknn-package.R
\docType{package}
\name{myknn-package}
\alias{myknn}
\alias{myknn-package}
\title{myknn: k-d Tree Nearest Neighbours}
\description{
Fast k-nearest neighbour search with a pure C k-d tree core and R bindings. Training data is referenced without copying from R matrices. Python bindings included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BartekDomanowski/myknn}
  \item Report bugs at \url{https://github.com/BartekDomanowski/myknn/issues}
}

}
\author{
\strong{Maintainer}: Bartłomiej Domanowski \email{01187317@pw.edu.pl} [copyright holder]

}
\keyword{internal}
