% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{route_cumulative_length_km}
\alias{route_cumulative_length_km}
\title{cumulative route length in array}
\usage{
route_cumulative_length_km(lats, lons)
}
\arguments{
\item{lats}{numeric vector of latitudes (degs).}

\item{lons}{numeric vector of longitudes (degs).}
}
\value{
numeric vector of cumulative distances in km.
}
\description{
returns cumulative haversine distance along a sequence of waypoints.
the first element is 0; the last equals \code{\link{route_length_km}(lats, lons)}.
}
