% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{track_cross_track_km}
\alias{track_cross_track_km}
\title{cross-track distance for each track point}
\usage{
track_cross_track_km(track_lats, track_lons, plan_lats, plan_lons)
}
\arguments{
\item{track_lats, track_lons}{actual track (degs).}

\item{plan_lats, plan_lons}{planned route (degs).}
}
\value{
numeric vector of cross-track distances in km (same length as track).
}
\description{
for each point on the track, returns cross-track distance to the
nearest segment of the planned route.
}
