% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirRoute-package.R
\docType{package}
\name{AirRoute-package}
\alias{AirRoute}
\alias{AirRoute-package}
\title{AirRoute: AirRoute - Aviation Route Metrics}
\description{
flight-route metrics, track deviation, airspace proximity detection, comparison with quitefastmst as a bonus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BartekDomanowski/AirRoute}
  \item Report bugs at \url{https://github.com/BartekDomanowski/AirRoute/issues}
}

}
\author{
\strong{Maintainer}: Bartłomiej Domanowski \email{01187317@pw.edu.pl} [copyright holder]

}
\keyword{internal}
